Added support for query arguments first, last, orderBy and offset
First is forward scrolling from offset if given.
Last is reverse scrolling from offset if given.
Default orderBy is the primary key. To order on multiple columns,
seperate the fields with ",". To optionally specify ascending or
descending order, prepend a "+" or "-" to the field. E.g.
"-first_name, last_name".