Added support for segmentation
First is forward scrolling from offset if given.
Last is reverse scrolling from offset if given.
Default order_by is the primary key. To order on multiple columns, separate the fields with ",". To optionally specify ascending or descending order, prepend a "+" or "-" to the field. E.g. "-first_name, last_name".